Move the submit wrapper back into 'formdata_listener_added.ts'.
