+++ packages/formdata-event/ts_src/wrappers/event_target.ts
@@ -12,9 +12,8 @@
 import {prototype as EventTargetPrototype, methods as EventTargetMethods} from '../environment/event_target.js';
 import {prototype as NodePrototype, methods as NodeMethods} from '../environment/node.js';
 import {prototype as WindowPrototype, methods as WindowMethods} from '../environment/window.js';
-import {formdataListenerAdded, formdataListenerRemoved} from '../formdata_listener_added.js';
+import {formdataListenerAdded, formdataListenerRemoved, wrapSubmitListener} from '../formdata_listener_added.js';
 import {submitListenerAdded, submitListenerRemoved} from '../submit_listener_added.js';
-import {wrapSubmitListener} from '../submit_listener_added.js';
 
 const submitListenerToWrapper = new WeakMap<EventListenerOrEventListenerObject, EventListener>();
